requirements.txt
flask>=2.3.0
moviepy>=1.0.3
opencv-python>=4.8.0
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
scipy>=1.11.0
scikit-image>=0.22.0
#tensorflow>=2.14.0
#keras>=2.14.0
flask-sqlalchemy>=3.1.0
sqlalchemy>=2.0.0
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
requests>=2.31.0 